requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
sqlalchemy==2.0.41
psycopg2==2.9.10
alembic==1.16.1
python-dotenv==1.1.0
pydantic==2.11.5
email-validator==2.2.0
pytest==8.3.5
pytest-asyncio==1.0.0
pytest-cov==6.1.1
httpx==0.28.1
passlib==1.7.4
python-jose==3.5.0
bcrypt==4.0.1
python-multipart==0.0.20